fix failed unit tests
