+++ parquet/src/util/mod.rs
@@ -21,6 +21,7 @@ mod bit_pack;
 pub(crate) mod interner;
 #[cfg(any(test, feature = "test_common"))]
 pub(crate) mod test_common;
+pub mod vec_util;
 
 #[cfg(any(test, feature = "test_common"))]
 pub use self::test_common::page_util::{
